+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetIntentConfiguration.swift
@@ -192,13 +192,13 @@ public extension PaymentSheet {
         /// Controls when the funds will be captured. 
         /// - Seealso: https://stripe.com/docs/api/payment_intents/create#create_payment_intent-capture_method
         public enum CaptureMethod: String {
-            /// (Default) Stripe automatically captures funds when the customer authorizes the payment.
+            /// Stripe automatically captures funds when the customer authorizes the payment.
             case automatic = "automatic"
 
             /// Place a hold on the funds when the customer authorizes the payment, but don’t capture the funds until later. (Not all payment methods support this.)
             case manual = "manual"
 
-            /// Asynchronously capture funds when the customer authorizes the payment.
+            /// (Default) Asynchronously capture funds when the customer authorizes the payment.
             /// - Note: Recommended over `CaptureMethod.automatic` due to improved latency, but may require additional integration changes.
             /// - Seealso: https://stripe.com/docs/payments/payment-intents/asynchronous-capture-automatic-async
             case automaticAsync = "automatic_async"
@@ -241,7 +241,7 @@ public extension PaymentSheet {
                 /// - Seealso: https://stripe.com/docs/api/payment_intents/create#create_payment_intent-setup_future_usage
                 setupFutureUsage: SetupFutureUsage? = nil,
                 /// - Seealso: https://stripe.com/docs/api/payment_intents/create#create_payment_intent-capture_method
-                captureMethod: CaptureMethod = .automatic,
+                captureMethod: CaptureMethod = .automaticAsync,
                 /// Additional payment method options params
                 /// - Seealso: https://docs.stripe.com/api/payment_intents/create#create_payment_intent-payment_method_options
                 paymentMethodOptions: PaymentMethodOptions? = nil